requirements.txt
numpy==1.24.3
opencv-python==4.7.0.72
Pillow==9.5.0
streamlit==1.22.0
tensorflow==2.10.1
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.31.0
